+++ packages/labnative/pages/index.js
@@ -13,6 +13,7 @@ import Input from './Input';
 import Progress from './Progress';
 import Dropdown from './Dropdown';
 import TextArea from './TextArea';
+import IconsList from './IconsList';
 import Icon from './Icon';
 
 export {
@@ -31,5 +32,6 @@ export {
   Progress,
   Dropdown,
   TextArea,
+  IconsList,
   Icon,
 };