+++ packages/yoga/src/Input/native/Input.jsx
@@ -168,9 +168,9 @@ const Input = ({
   textContentType,
   value,
   onBlur,
-  onChangeText,
   onClean,
   onFocus,
+  hideMaxLength,
   theme: {
     yoga: {
       colors,
@@ -181,11 +181,9 @@ const Input = ({
   ...props
 }) => {
   const [focused, setFocused] = useState(false);
-  const [inputValue, setInputValue] = useState(value);
   const [typed, setTyped] = useState(Boolean(value));
 
   useEffect(() => {
-    setInputValue(value);
     setTyped(Boolean(value));
   }, [value]);
 
@@ -246,12 +244,7 @@ const Input = ({
         }}
         textContentType={textContentType}
         typed={typed}
-        value={inputValue}
-        onChangeText={text => {
-          setTyped(Boolean(text));
-          setInputValue(text);
-          onChangeText(text);
-        }}
+        value={value}
         onFocus={e => {
           setFocused(true);
           onFocus(e);
@@ -300,7 +293,8 @@ const Input = ({
           focused={focused}
           disabled={disabled}
           maxLength={maxLength}
-          length={inputValue.length}
+          length={value.length}
+          hideMaxLength={hideMaxLength}
         />
       )}
     </Wrapper>
@@ -323,10 +317,10 @@ Input.propTypes = {
   value: oneOfType([string, number]),
   style: oneOfType([shape({}), arrayOf(shape({}))]),
   onBlur: func,
-  onChangeText: func,
   /** callback invoked when close icon is clicked */
   onClean: func,
   onFocus: func,
+  hideMaxLength: bool,
 };
 
 Input.defaultProps = {
@@ -342,9 +336,9 @@ Input.defaultProps = {
   textContentType: undefined,
   value: '',
   onBlur: () => {},
-  onChangeText: () => {},
   onClean: () => {},
   onFocus: () => {},
+  hideMaxLength: false,
 };
 
 export default withTheme(Input);