+++ packages/yoga/src/Input/web/Fieldset.jsx
@@ -6,6 +6,8 @@ const Fieldset = styled.fieldset`
   margin: 0;
   padding: 0;
 
+  box-sizing: border-box;
+
   ${({
     error,
     disabled,