+++ packages/yoga/src/index.native.js
@@ -13,6 +13,7 @@ import Input from './Input';
 import Progress from './Progress';
 import Dropdown from './Dropdown';
 import TextArea from './TextArea';
+import Icon from './Icon';
 
 export {
   ThemeProvider,
@@ -35,4 +36,5 @@ export {
   Progress,
   Dropdown,
   TextArea,
+  Icon,
 };