+++ packages/labnative/CHANGELOG.md
@@ -3,37 +3,63 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## 0.16.9-alpha.4 (2021-01-27)
+# [0.17.0-alpha.3](https://github.com/Gympass/yoga/compare/@gympass/yoga-labnative@0.17.0-alpha.2...@gympass/yoga-labnative@0.17.0-alpha.3) (2021-02-17)
 
 **Note:** Version bump only for package @gympass/yoga-labnative
 
+## [0.16.13](https://github.com/Gympass/yoga/compare/@gympass/yoga-labnative@0.16.12...@gympass/yoga-labnative@0.16.13) (2021-02-09)
 
+**Note:** Version bump only for package @gympass/yoga-labnative
 
+## [0.16.12](https://github.com/Gympass/yoga/compare/@gympass/yoga-labnative@0.16.11...@gympass/yoga-labnative@0.16.12) (2020-12-16)
 
+**Note:** Version bump only for package @gympass/yoga-labnative
+
+## [0.16.11](https://github.com/Gympass/yoga/compare/@gympass/yoga-labnative@0.16.10...@gympass/yoga-labnative@0.16.11) (2020-12-11)
+
+**Note:** Version bump only for package @gympass/yoga-labnative
+
+## [0.16.10](https://github.com/Gympass/yoga/compare/@gympass/yoga-labnative@0.16.9...@gympass/yoga-labnative@0.16.10) (2020-12-10)
+
+**Note:** Version bump only for package @gympass/yoga-labnative
+
+## [0.16.9](https://github.com/Gympass/yoga/compare/@gympass/yoga-labnative@0.16.8...@gympass/yoga-labnative@0.16.9) (2020-11-26)
+
+**Note:** Version bump only for package @gympass/yoga-labnative
+
+# [0.17.0-alpha.2](https://github.com/Gympass/yoga/compare/@gympass/yoga-labnative@0.17.0-alpha.1...@gympass/yoga-labnative@0.17.0-alpha.2) (2021-02-12)
+
+**Note:** Version bump only for package @gympass/yoga-labnative
+
+# [0.17.0-alpha.1](https://github.com/Gympass/yoga/compare/@gympass/yoga-labnative@0.17.0-alpha.0...@gympass/yoga-labnative@0.17.0-alpha.1) (2021-02-11)
+
+**Note:** Version bump only for package @gympass/yoga-labnative
+
+# [0.17.0-alpha.0](https://github.com/Gympass/yoga/compare/@gympass/yoga-labnative@0.16.9-alpha.5...@gympass/yoga-labnative@0.17.0-alpha.0) (2021-02-11)
+
+### Features
+
+- **iconwrapper:** add Icon wrapper abstraction component ([a1c19ea](https://github.com/Gympass/yoga/commit/a1c19ead37f8e592101cf89bcbd749d9e491533a))
+
+## [0.16.9-alpha.5](https://github.com/Gympass/yoga/compare/@gympass/yoga-labnative@0.16.9-alpha.4...@gympass/yoga-labnative@0.16.9-alpha.5) (2021-02-04)
+
+**Note:** Version bump only for package @gympass/yoga-labnative
+
+## 0.16.9-alpha.4 (2021-01-27)
+
+**Note:** Version bump only for package @gympass/yoga-labnative
 
 ## 0.16.9-alpha.3 (2021-01-22)
 
 **Note:** Version bump only for package @gympass/yoga-labnative
 
-
-
-
-
 ## 0.16.9-alpha.2 (2020-12-03)
 
 **Note:** Version bump only for package @gympass/yoga-labnative
 
-
-
-
-
 ## 0.16.9-alpha.1 (2020-10-28)
 
 **Note:** Version bump only for package @gympass/yoga-labnative
-
-
-
-
 
 ## 0.16.9-alpha.0 (2020-10-23)